When in read-only remove a tear in the character below the cursor position.
What happens is that when the timer fires, viewport()->update(_cursorRect); is called.
This has the side effect to erase the cursor rectangle and the paint event must repaint it.
Problem is that the rectanlge is in the QRect of the character *below* which is *not* repainted.
(Use a super tall character like A with some accent (C3)).

There is no need to fire the timer in read-only mode so this problem does not happen any longer.

It is still visible though in !readOnly especially if the overwrite mode is on (vertical curson).
One can see that when the cursor is *not* blinking, the space behind is white,
 while it should contain part of the character.
Most character will show this issue.


Signed-off-by: Andrea Odetti <mariofutire@gmail.com>